Enable Glimr API case fee requests (#6)
* Enable Glimr API case fee requests

* Use consistent config and setup test environment

HTTParty loads the base_uri setting too early to allow us to reset or
mock it using standard methods.  After discussions, Christian and I
decided to use a committed copy of `.env.test` to deal with this
problem: it offers a solution to the problem that is consistent to the
way we run the app in development mode and does not tie the solution to
any particular CI.

* Enable Glimr API fee paid requests

* Enable Glimr API case fee responses

* Keep circle build consistent with development

The environment variables were not all being picked up when set in the
`circle.yml` file.  However, setting them by copying `env.circle` to
`.env.test` after checkout, but before the test run solved the problem.