Continually check that glimr is available
Previously, we were only checking this at the start of the run. This did
not account for the possibility that glimr could go away at any point in
the 'transaction'. The user would not have been notified until they
attempted to complete a payment.  This will stop them as move between
'states'/pages.