Rename fee_liabilities to liabilities
Both were being used, so I opted for the shorter one.  I though about
'fees' as well, but it wasn’t being used anywhere.  'fee_liabilities' is
what glimr uses, and, again, I though about tracking that.  However, it
only gets called in a couple of places specifically for the API, so I
decided to save on the typing, overall.