Remove govpay cruft (#14)
* Make the logfiles easier to digest

Even with lograge, it was still dumping all the activerecord requests
into the logs.  This made them difficult to read/follow.

* Make the tests easier to maintain and deal with

The unit tests were (predictably) fragile.  The request specs give
better coverage and a clearer indication of what is going on.

* Store all the settings in the same place

* Stop govpay from choking

It doesn't like redirects to localhost:3000. This is what Chirstian used
in the prototype and works fine. You do have to change the protocol and
url after it returns, however.

* Prevent govpay errors from bubbling up

This alerts the user in a sensible way (with placeholder copy).