Simplify the case request workflow
After discussion with David, it seemed sensible to remove the
consolidator forms/new_case_request logic into case_request, itself.
This reduced the complexity, particularly in the case_request
controller.  It also makes it more straightforward to look up existing
case requests and their associated liabilities.

TODO: Now that we store the case confirmation, would it be sensible to
encrypt it in our store object?