Save work showing the current state
I am making this branch so we can revisit the problem later. Right now,
I do not have time to investigate exactly where/why it is occurring.

Summary:  Webmock stubs do not seem to get added to the registered stubs
if they are declared in a `describe`, `context`, or `feature` block that
is itself declared inside the initial RSpec.feature declaration. This
only happens with feature specs.