Already paid liabilities (#15)
* Rename fee_liabilities to liabilities

Both were being used, so I opted for the shorter one.  I though about
'fees' as well, but it wasn’t being used anywhere.  'fee_liabilities' is
what glimr uses, and, again, I though about tracking that.  However, it
only gets called in a couple of places specifically for the API, so I
decided to save on the typing, overall.

* Find existing case requests first

This is a precursor to being able to identify liabilities that have
already been paid.

* Reduce the ambiguity of the payment process

We know when a user has paid.  Now we can tell them they have paid and
advise them that other fees may be due later, and that the tribunal will
let them know if this is the case.

TODO:

* Spec the case where additional fees become due.
* Get proper wording and design in place for the notification
* Require confirmation code for local case requests

* Bring specs into alignment

* Simplify the case request workflow

After discussion with David, it seemed sensible to remove the
consolidator forms/new_case_request logic into case_request, itself.
This reduced the complexity, particularly in the case_request
controller.  It also makes it more straightforward to look up existing
case requests and their associated liabilities.

TODO: Now that we store the case confirmation, would it be sensible to
encrypt it in our store object?

* move factory_girl and database_cleaner to test group

Now that lib/tasks/factory_girl.rake has been deleted, there is no
longer any need for these gems to be in the development group.