Rename liabilites and disconnect case_requests
DS and I discussed the various race conditions I was encountering when
creating and persisting `CaseRequest` models. We decided that the
best way to handle these problems, which were numerous and sometimes
subtle, was to dispense with persisting case requests and only store
the fee information we get back from glimr in `Fee` models.

A `CaseRequest` model still exists, but it has been refactored into a
transient convenience container for the collection of methods used to
look up a case and get its information from glimr.  `Liability` has been
renamed again to `Fee` for brevity.  The `confirmation_code` is now
stored as a `BCrypt` digest on each `Fee` object.  Ultimately, we may
not need the code at all, or it may be OK to store it in plaintext.  For
now, however, I decided it was better to deal with the (minor)
complexity of encrypting it as part of the storage process.