Simplify and clarify the glimr api client requests
HTTParty can sometimes feel a little too magic.  I have replaced it with
RestClient, which is normally called directly.  This simplified the
workflow. I also moved things from the inherited Base class into the Api
class and mixed them in.  Finally, after discussing it with DS, I
removed the multi-level class structure from the response trees for a
couple of the calls.  The calls in question (availability, for example),
expressed simple results and did not need the level of interface
consistency expressed by the previous implementation.