Glimr as source of truth (#16)
* Remove virtus

We don’t need this any longer.

* Rename liabilites and disconnect case_requests

DS and I discussed the various race conditions I was encountering when
creating and persisting `CaseRequest` models. We decided that the
best way to handle these problems, which were numerous and sometimes
subtle, was to dispense with persisting case requests and only store
the fee information we get back from glimr in `Fee` models.

A `CaseRequest` model still exists, but it has been refactored into a
transient convenience container for the collection of methods used to
look up a case and get its information from glimr.  `Liability` has been
renamed again to `Fee` for brevity.  The `confirmation_code` is now
stored as a `BCrypt` digest on each `Fee` object.  Ultimately, we may
not need the code at all, or it may be OK to store it in plaintext.  For
now, however, I decided it was better to deal with the (minor)
complexity of encrypting it as part of the storage process.

* Move CaseRequest to a more sensible category

While is does have many of the trappings of a model, the real role of
`CaseRequest` is to aggregate a collection of data returned by the glimr
api.  Currently, this is limited to fee requests, but will probably
include case documents sometime in the near future.

* Change liability to fee, Spec multiple fees

We were using fee_liability and liability inconsistentely. I settled on
‘fee’ as the simplest possible nomeclature. It is a little inconsistent
with glimr, which seems to use `feeLiabilities`.  I think is is unlikely
that anyone will become confused by the translation, and not having to
constantly type 'fee_liability' feels like it worth any possible
confusion.

I have also simplified some of the glimr shared examples, renaming them
according to what they are doing.  Previously, they had abstract-ish
names and expected data structures as arguments. The arguments were
almost always the same, so I refactored for clarity and brevity.