Use glimr as the source of truth for case requests
Previously, there were multiple ways to validate a case request. It
could be validated:

  * locally, based on the information submitted by the user;
  * by glimr returning a 404;
  * hypothetically, by glimr returning a JSON response indicating
the case could not be found (glimr does not currently do this);

Each of these methods took a different path through the app. It was
possible for the same error to be raised in response to different
conditions. This was confusing.

This commit makes glimr the sole source of truth; It only accounts for
the case where glimr returns a 404 when a case cannot be found.