prevent simplecov/mutant conflicts
Mutant tests a single class at at time. simplecov
interprets this as low coverage (because the tests
for a single class don't cover the whole project).
This change enables mutant to run the specs
without triggering simplecov, so circleci can use 
the exit status of the mutant run to decide
whether the build is successful or not.