Refactor glimr api (#18)
* Make the category more explicit

‘service’ felt somewhat too general as both the glimr and govpay code
are api_clients.

* Make the simpler option available

HTTParty can be difficult to mock and sometimes feels a bit too ‘magic’.
Rest-client is simpler and arguably the most widely-used gem for the
purpose.

I am leaving HTTParty in the Gemfile for now as the govpay client still
uses it.

* Simplify and clarify the glimr api client requests

HTTParty can sometimes feel a little too magic.  I have replaced it with
RestClient, which is normally called directly.  This simplified the
workflow. I also moved things from the inherited Base class into the Api
class and mixed them in.  Finally, after discussing it with DS, I
removed the multi-level class structure from the response trees for a
couple of the calls.  The calls in question (availability, for example),
expressed simple results and did not need the level of interface
consistency expressed by the previous implementation.

* Use glimr as the source of truth for case requests

Previously, there were multiple ways to validate a case request. It
could be validated:

  * locally, based on the information submitted by the user;
  * by glimr returning a 404;
  * hypothetically, by glimr returning a JSON response indicating
the case could not be found (glimr does not currently do this);

Each of these methods took a different path through the app. It was
possible for the same error to be raised in response to different
conditions. This was confusing.

This commit makes glimr the sole source of truth; It only accounts for
the case where glimr returns a 404 when a case cannot be found.