add mutation testing (#19)
* add mutation testing

Add mutant-rspec gem and a "mutant" rake task,
plus specs and code changes to get to 100%
mutation testing coverage on the CaseRequest model

TODO: Add a better runner to exercise all relevant
code and bring coverage up to 100% throughout.

* case reference & conf. code as keyword arguments

Instead of passing a hash to CaseRequest.new, explicitly provide
case_reference and confirmation_code

* check for error in process!

If the case was not found, then we have a CaseNotFound object,
which doesn't implement "fee_liabilities."

* fix rubocop's complaints

* extract mutant to lib/tasks/mutant.rake

* prevent simplecov/mutant conflicts

Mutant tests a single class at at time. simplecov
interprets this as low coverage (because the tests
for a single class don't cover the whole project).
This change enables mutant to run the specs
without triggering simplecov, so circleci can use 
the exit status of the mutant run to decide
whether the build is successful or not.

* add mutation testing

Add mutant-rspec gem and a "mutant" rake task,
plus specs and code changes to get to 100%
mutation testing coverage on the CaseRequest model

TODO: Add a better runner to exercise all relevant
code and bring coverage up to 100% throughout.

* case reference & conf. code as keyword arguments

Instead of passing a hash to CaseRequest.new, explicitly provide
case_reference and confirmation_code

* check for error in process!

If the case was not found, then we have a CaseNotFound object,
which doesn't implement "fee_liabilities."

* fix rubocop's complaints

* extract mutant to lib/tasks/mutant.rake

* prevent simplecov/mutant conflicts

Mutant tests a single class at at time. simplecov
interprets this as low coverage (because the tests
for a single class don't cover the whole project).
This change enables mutant to run the specs
without triggering simplecov, so circleci can use 
the exit status of the mutant run to decide
whether the build is successful or not.

* integrate latest changes from master

* fix alphabetical ordering of gems

* use house alignment style for '=' and '{}'

* really fix brace alignment, this time.