Replace webmock/rest-client with excon
Changing expected headers in the govpay api webmock stubs casued
rest-client to send different headers when running the specs. This was
worrying and neither DS nor I could find an obvious cause.  I decided to
change out to Excon, which is has its own mocking, as it is simpler and
being used successfully by Prison Visit Boookings. This branch is a
refactor of the already-completed glimr api refactor to achieve this.

Webmock and rest-client are being left in as they are still required by
the govpay api refactor feature branch.  I will remove them as part of
the changes on that branch.