Add mutation testing of fee model (#21)
* WiP: starting on Fee specs

* add mutation testing for Fee model

* force timezone to somewhere without summer time

This (should) make the tests break when mutant changes
'Time.zone.now' to 'Time.now', when the CI server runs mutant.

* remove blank lines to keep rubocop happy

* use built-in rails time helpers instead of timecop

* suppress warning message if module is reloaded

This occurs during mutation testing.

* mutation test all models

This change means it's no longer necessary to list out the model
classes that should be mutation tested. Any new model class in the
app/models directory, whether it's an ActiveRecord class or a PORO,
should be mutation tested as part of our build process.