Simplify the workflow around errors
There are several workflows a govpay error might take.  The aim of this
commit is to simplify and expose them somewhat.  Previously, they had
mainly been encapsulated in response objects.  Now all errors that are
not users errors in completing the govpay credit card payment (e.g. a
bad CVV number) *should* be raised as exceptions and rescued in the
FeesController.