Refactor govpay api (#22)
* Simplify govpay api client, convert to excon

This changes the govpay api to follow the excon changes that were
previously implemented on the glimr api. It allows a cleaner workflow
through the app.

* Simplify the workflow around errors

There are several workflows a govpay error might take.  The aim of this
commit is to simplify and expose them somewhat.  Previously, they had
mainly been encapsulated in response objects.  Now all errors that are
not users errors in completing the govpay credit card payment (e.g. a
bad CVV number) *should* be raised as exceptions and rescued in the
FeesController.

* Ensure that api data errors are clear

It was not always clear why the app failed when a partial message was
received. This makes it much more explicit and exposed the fact that
message is only used for errors (see
`govpay/resposnes/payment_status.rb` for details).