Refactor specs to make them clearer and easier
The request specs are harder to diagnose/maintain than the feature
specs.  Furthermore, a request spec is only required when a payment is
first created.  This is because the app needs to redirect to the Gov UK
Payment gateway, and a feature spec would be unable to follow this
redirection.  As a result, I have refactored all but one of the create
payment specs from request specs into feature specs.  While I was doing
this, I removed some lint from the shared examples and DRY'd out a
couple of things, like the Fee object creation logic.  See the code for
further notes.