Raise test exceptions cleanly
When I started working with Excon, I hadn’t realised it was possible to
have it return a block that raised an exception.  This is much more
straightforward than trying to stub out the class and set expectations
on it.  These changes reflect my updated understanding of how it works.