Use govuk pay gem (#25)
* Use an external gem for GovUK Payment interactions

This commit uses the GovUK Payment gateway client gem instead of the
internal client. This was done for ease-of-use and to simplify
reusability and maintainability.

* Refactor specs to make them clearer and easier

The request specs are harder to diagnose/maintain than the feature
specs.  Furthermore, a request spec is only required when a payment is
first created.  This is because the app needs to redirect to the Gov UK
Payment gateway, and a feature spec would be unable to follow this
redirection.  As a result, I have refactored all but one of the create
payment specs from request specs into feature specs.  While I was doing
this, I removed some lint from the shared examples and DRY'd out a
couple of things, like the Fee object creation logic.  See the code for
further notes.

* Update the version of the govpay gem

The previous version had a rake task that allowed the end user to
install a set of RSpec shared examples. This task *appeared* to have
been run every time a full spec run was made.  Therefore, any changes
made on the shared examples would wind up getting automatically (and
silently) overwritten after each full spec run.  This fixes that.

* Raise test exceptions cleanly

When I started working with Excon, I hadn’t realised it was possible to
have it return a block that raised an exception.  This is much more
straightforward than trying to stub out the class and set expectations
on it.  These changes reflect my updated understanding of how it works.

* Change from expections to actual specs

Or, maybe better said, specs that check the return values of the model
attributes after encryption/saving.  Ultimately seems cleaner then
mocking and doesn’t stray too far from the model, proper.

* Clairify the purpose of the attribute

It isn’t a payment url, but rather the URL that the GovUK Pay gateway
*should* return a user to after a successful or attempted payment.

* Do not overwrite glimr shared examples

This was having the same problem that the GovUK Payment gateway gem was
having; the shared examples were being overwritten on every full RSpec
run.  This is fixed in the GLiMR gem from version 0.1.3.

* Shut rubocop up about differing use cases

This particular cop just really isn’t very helpful.

* Clean up various rubocop issues

* Include a mirror spec to clairifiy negative specs

DS pointed out that the RSpec expectations ensure that GLiMR updates
were not being called were not sufficiently contextualised. Specs
existed to ensure positive cases were called, but these used Excon stubs
instead of RSpec stubs.  In short, their relation to the negative specs
was not explicit.  This DRYs out the negative specs into shared examples
and adds a positive shared spec, which should clairify the intent of the
negative specs.