Stop hidden api calls from blocking specs
One of the specs was call glimr in the background. As I was WIP'd, the
failures on Circle were not top priority.  I only realised what was
happening when I worked from home one day and the one of the specs
started timing out.  While the intent is always to mock out the api
calls, proper, we still need to leave webmock in place to ensure that
no new ones slip by unnoticed.