PaymentsController#show to handle success messages
This was being dealt with by FeesController, but that is no longer the
best place for it.  Now, the payment method controller will handle
errors, while PaymentsController will handle successes. This is because
the errors, when they are surfaced, are more specific to the call than
the successes messages–which are just text.