Allow payment via help with fees reference (#35)
* Expose separate route for help with fees payments

* More work on help with fees

I've just realised I need the full workflow to avoid test smells. This
commit is the stopping place for the help with fees method while I work
out the routing to the specific payment methods.

* Bring payment page in line with prototype

This includes the layout changes and selecting the payment
type–including help with fees.

TODO: Update the styles to use the up-to-date version of the frontend
toolkit.

* Update the govuk_frontend toolkit

* Make the alpha banner easier to change

* Move the main content into the left column

* Populate right hand column

* Move the styles over from the prototype

This will need some additional tweaking, but it gets the app very close
to the prototype with minimal effort.

* Add placeholder for step numbering

* Update the wording to agree with prototype

* Add missing class

This is in the prototype, but not here.

* Get rid of doubled liabilities border

* Fix the missing icon

Rails does *not* require the path to be set.

* Note on status of GOVUK elements

* Offer technical help to users

* Put all the copy in the same place

Make it easier to maintain.

* Bring the payment form inline with the prototype

This follow the basic layout of the prototype, but will need some
additional work on the styling (have already discussed with Clive).

* Bring the start page in line with the prototype

* Put the case not found message where it belongs

It should not have been bouncing users all the way to the start page.

* Make the payment methods logic cleaner

This was being handled in CaseRequestsController.  However, after
discussing the WIP with David, it seemed more sensible to move it into
it's own controller.

* Route via the help with fees controller

This is a step towards getting everything into specific controllers that
deal with the individual payment types.

* Fix the styling on the button

* Stop hidden api calls from blocking specs

One of the specs was call glimr in the background. As I was WIP'd, the
failures on Circle were not top priority.  I only realised what was
happening when I worked from home one day and the one of the specs
started timing out.  While the intent is always to mock out the api
calls, proper, we still need to leave webmock in place to ensure that
no new ones slip by unnoticed.

* Stop failure due to unstubbed GLiMR call

I missed the before block when I was setting this up.  It was working on
the MoJ network because GLiMR was responding as expected.

* Broaden scope of payment methods controller

Renamed it to PaymentsController as it is now going to be the final page
to show the result of all successful payments.

* Allow help with fees to be tracked from fee object

* Separate out help with fees concerns further

It was using the payments layout, which is now just used to show the
payment results pages.  Plus neaten up the html.

* Process help with fees request and show response

This is also the starting place from making `Payments#show` the main
page for dealing with all payment response messages.

* PaymentsController#show to handle success messages

This was being dealt with by FeesController, but that is no longer the
best place for it.  Now, the payment method controller will handle
errors, while PaymentsController will handle successes. This is because
the errors, when they are surfaced, are more specific to the call than
the successes messages–which are just text.

* Fix coverage, style and security issues

Nothing major, just irrelevant error handling/logging a very minor XSS
risk and a known rubocop issue.

* Stop glimr update call from breaking

Originally, the api call was expecting a fee object.  This changed a
while back, but it was missed in this app until we started trying to
integrate the stack in a live(ish) environment.

* Bring the wording inline with the standard vocab

* Fix typo

* Remove unneeded stylesheets

As per @csutter's comments on the PR