Prep work to stop making unneeded fee objects
With the introduction of the API call to create a fee payment request,
the current design of the system made it necessary to call
`CaseRequest#process!` in multiple actions.  This was because there
was no association between `CaseRequest`–which has only started to be
persisted as part of this PR–and `Fee`.  The problem comes in when a
newly-created `CaseRequest` is shown; in order to determine its fee
liabilities, `CaseRequest#process!` has to be called.  However, this
creates new `Fee` objects, which is inappropriate for a `show` action.

By adding the correct `has_many` association, the fees can be
determined and persisted in `#create`, leaving `#show` to, well, show...

I originally chose not to do this as it was more work owing to the fact
that `CaseRequest#fees`, which *should* be the name of the association,
is actually a method delegated to the GLiMR API client.  I have now
worke around that by naming the `Fee` has_many association `#case_fees`.