Datacapture integration (#38)
* Set up mime response blocks

Initial step in getting it to respond to JSON requests.

* Recreate case request table

This was removed a while back, but is required again as it is the
simplest way of persisting a CaseRequest for the API calls.

* Clear the database between runs

This got overlooked some time ago, so records were hangning around
between test runs.

* Send Datacapture app the redirect url

This is happy path for delivery of the url to the datacapture app. It
does not yet encompass edge cases nor the case where the datacapture app
sends the user back for payment.

* Notify datacapture when the case is not found

This is the 'happy path' failure case–i.e. GLiMR cannot find the case
reference.

* Handle the common error cases

* Process and show the case request

This has been broken out into its own action in order to integrate it
with the API call that will be used by the data capture app.

* Redirect users to new if a case is not found

* Prep work to stop making unneeded fee objects

With the introduction of the API call to create a fee payment request,
the current design of the system made it necessary to call
`CaseRequest#process!` in multiple actions.  This was because there
was no association between `CaseRequest`–which has only started to be
persisted as part of this PR–and `Fee`.  The problem comes in when a
newly-created `CaseRequest` is shown; in order to determine its fee
liabilities, `CaseRequest#process!` has to be called.  However, this
creates new `Fee` objects, which is inappropriate for a `show` action.

By adding the correct `has_many` association, the fees can be
determined and persisted in `#create`, leaving `#show` to, well, show...

I originally chose not to do this as it was more work owing to the fact
that `CaseRequest#fees`, which *should* be the name of the association,
is actually a method delegated to the GLiMR API client.  I have now
worke around that by naming the `Fee` has_many association `#case_fees`.

* Reduce the level of surprise

As per the previous commit, this was a (smelly) workaround. It is no
longer needed now that the associations are in places and behaving as
expected.