Log all failures to find the case
`#show` fails softly and redirects to `#new`.  I added logging to the
action to ensure we can tell what is going on.  The `302` that otherwise
appears in the log is probably not sufficient if we should need to
analyse problems.

`SimplifiedLogging#log_error` can now also take an arbitrary number of
additional key-value pairs to append to the log message.