Log pay by account errors
These would not have been clearly logged.  Also, the error handling
code that was in place (which was copy-pasted from an older version of
the API client) was not relevant in this context, so I removed it.

It is worth noting here that some errors bubble up and are re-rescued by
the controller.  This is done so the user can be notified they are
entering incorrect information.