Log all glimr calls (#40)
* Log all failures to find the case

`#show` fails softly and redirects to `#new`.  I added logging to the
action to ensure we can tell what is going on.  The `302` that otherwise
appears in the log is probably not sufficient if we should need to
analyse problems.

`SimplifiedLogging#log_error` can now also take an arbitrary number of
additional key-value pairs to append to the log message.

* fixup! Log all failures to find the case

* Log help with fees errors

These would not have been clearly logged.  Also, the error handling code
that was in place (which was copy-pasted from an older version of the
API client) was not relevant in this context, so I removed it.

* Log pay by account errors

These would not have been clearly logged.  Also, the error handling
code that was in place (which was copy-pasted from an older version of
the API client) was not relevant in this context, so I removed it.

It is worth noting here that some errors bubble up and are re-rescued by
the controller.  This is done so the user can be notified they are
entering incorrect information.