Implement basic healthcheck
At this time it does not check govuk pay for status.  When the govuk pay
gem was written, govuk pay did not have a status endpoint.  I need to
check if one has been added since. If so, it will require changes to
the govuk pay gem before it can be implemented.  In the interest of
getting this released in an MVP state, I have de-scoped that for now.