Enable govuk pay healthcheck
I was going to leave this out as I could not find documentation for it
and no one responded when I asked on govuk_pay slack. However, I
discovered the endpoint by trial and error and decided to put in a basic
check based on what it is currently returning.  For now I am not going
to port this to the govuk_pay gem as I do not have documentation, but
that should be on the todo list for a future release.