Healthcheck (#47)
* Correct oversight

This should have been random from the start.

* Implement basic healthcheck

At this time it does not check govuk pay for status.  When the govuk pay
gem was written, govuk pay did not have a status endpoint.  I need to
check if one has been added since. If so, it will require changes to
the govuk pay gem before it can be implemented.  In the interest of
getting this released in an MVP state, I have de-scoped that for now.

* Enable govuk pay healthcheck

I was going to leave this out as I could not find documentation for it
and no one responded when I asked on govuk_pay slack. However, I
discovered the endpoint by trial and error and decided to put in a basic
check based on what it is currently returning.  For now I am not going
to port this to the govuk_pay gem as I do not have documentation, but
that should be on the todo list for a future release.